Fixing a failing unit test
