+++ api/pom.xml
@@ -100,6 +100,11 @@
             <version>4.13.1</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>de.lars-sh</groupId>
+            <artifactId>lombok-annotations</artifactId>
+            <version>1.18.16</version>
+        </dependency>
 
     </dependencies>
